# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Forum API",
    "description": "Essa API é composta por endpoints que recebem e respondem a solicitações de clientes usando um protocolo de comunicação comum, geralmente HTTP. Esses endpoints fornecem acesso aos recursos do fórum, como usuários, categorias, comentários e outras informações relevantes.",
    "contact": {
      "name": "Support",
      "email": "victoralmeida.0216@gmail.com"
    },
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "Usuários",
      "description": "Tudo sobre nossos usuários"
    },
    {
      "name": "Categorias",
      "description": "Tudo sobre as categorias"
    },
    {
      "name": "Comentários",
      "description": "Tudo sobre os comentários realizados no site"
    }
  ],
  "paths": {
    "/api/users": {
      "post": {
        "tags": ["Usuários"],
        "description": "Criação de um usuário",
        "requestBody": {
          "description": "Dados necessários para cadastro de um usuário",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cadastro realizado com sucesso!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseUser"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados incompletos!"
          }
        }
      },
      "get": {
        "description": "Usuários cadastrados",
        "tags": ["Usuários"],
        "responses": {
          "200": {
            "description": "Usuários retornados com sucesso!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUser"
                }
              }
            }
          },
          "400": {
            "description": "Erro!"
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "description": "Login de um usuário",
        "tags": ["Usuários"],
        "requestBody": {
          "description": "Dados necessários para o login de um usuário",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseLogin"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados incompletos!"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "description": "Busca de um usuário por id",
        "tags": ["Usuários"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do usuário",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário retornado com sucesso!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUser"
                }
              }
            }
          },
          "400": {
            "description": "Usuário não existe!"
          }
        }
      },
      "patch": {
        "description": "Atualização do usuário pelo id",
        "tags": ["Usuários"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do usuário",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "name_user",
            "in": "path",
            "description": "Nome do usuário",
            "required": false,
            "type": "string"
          },
          {
            "name": "email",
            "in": "path",
            "description": "Email do usuário",
            "required": false,
            "type": "string"
          },
          {
            "name": "password",
            "in": "path",
            "description": "Senha do usuário",
            "required": false,
            "type": "string"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dados atulizados com sucesso!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/ResponseUser"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados incompletos!"
          }
        }
      },
      "delete": {
        "description": "Deletar usuário pelo id",
        "tags": ["Usuários"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do usuário",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário deletado com sucesso!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUser"
                }
              }
            }
          },
          "400": {
            "description": "Usuário não existe!"
          }
        }
      }
    },
    "/api/categories": {
      "post": {
        "description": "Criação de uma categoria",
        "tags": ["Categorias"],
        "requestBody": {
          "description": "Dados necessários realizar o cadastro da categoria",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCategory"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "description": "Busca de todas as categorias",
        "tags": ["Categorias"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Categorias cadastradas!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseCategory"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro!"
          }
        }
      }
    },
    "/api/categories/{name}": {
      "get": {
        "description": "Busca de categoria pelo nome",
        "tags": ["Categorias"],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Nome da categoria",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Category: Java!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseCategory"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados incompletos!"
          }
        }
      }
    },
    "/api/categories/{id}": {
      "patch": {
        "description": "Atualizar categoria pelo ID",
        "tags": ["Categorias"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da categoria",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "name_category",
            "in": "path",
            "description": "Name da categoria",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Categoria atualizada sucesso!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseCategory"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados incompletos!"
          }
        }
      }
    },
    "/api/comments": {
      "post": {
        "tags": ["Comentários"],
        "description": "Criação de um comentário",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Dados necessários para cadastro de um comentário",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestComment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cadastro realizado com sucesso!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseComment"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados incompletos!"
          }
        }
      },
      "get": {
        "description": "Comentários cadastrados",
        "tags": ["Comentários"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Comentários criados!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseComment"
                }
              }
            }
          },
          "400": {
            "description": "Erro!"
          }
        }
      }
    },
    "/api/comments/categories/{id}": {
      "get": {
        "description": "Busca de todas as categorias",
        "tags": ["Comentários"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da categoria",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Categorias cadastradas!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseCategory"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro!"
          }
        }
      }
    },
    "/api/comments/users/{id}": {
      "get": {
        "description": "Busca de todos os usuários",
        "tags": ["Comentários"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do usuário",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Usuários cadastrados!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseUser"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro!"
          }
        }
      }
    },
    "/api/comments/{id}": {
      "patch": {
        "description": "Atualizar comentário pelo ID",
        "tags": ["Comentários"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da categoria",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "text_comment",
            "in": "path",
            "description": "Comentário",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Comentário atualizado sucesso!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseComment"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados incompletos!"
          }
        }
      },
      "delete": {
        "description": "Deletar comentário pelo ID",
        "tags": ["Comentários"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do comentário",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Comentário deletado sucesso!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseComment"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados incompletos!"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RequestLogin": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email do usuário a ser criado",
            "example": "joao@gmail.com"
          },
          "password": {
            "type": "string",
            "description": "Senha do usuário a ser criado",
            "example": "12345"
          }
        }
      },
      "ResponseLogin": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email do usuário a ser criado",
            "example": "joao@gmail.com"
          },
          "password": {
            "type": "string",
            "description": "Senha do usuário a ser criado",
            "example": "12345"
          },
          "token": {
            "type": "string",
            "description": "Token de autenticação"
          }
        }
      },
      "RequestUser": {
        "type": "object",
        "properties": {
          "name_user": {
            "type": "string",
            "description": "Nome do usuário a ser criado",
            "example": "João"
          },
          "email": {
            "type": "string",
            "description": "Email do usuário a ser criado",
            "example": "joao@gmail.com"
          },
          "password": {
            "type": "string",
            "description": "Senha do usuário a ser criado",
            "example": "12345"
          }
        }
      },
      "ResponseUser": {
        "type": "object",
        "properties": {
          "name_user": {
            "type": "string",
            "description": "Nome do usuário a ser criado",
            "example": "João"
          },
          "email": {
            "type": "string",
            "description": "Email do usuário a ser criado",
            "example": "joao@gmail.com"
          },
          "password": {
            "type": "string",
            "description": "Senha do usuário a ser criado",
            "example": "12345"
          },
          "id": {
            "type": "string",
            "description": "Id do usuário",
            "format": "uuid"
          }
        }
      },
      "RequestCategory": {
        "type": "object",
        "properties": {
          "name_category": {
            "type": "string",
            "description": "Nome da categoria a ser criada",
            "example": "Prisma"
          }
        }
      },
      "ResponseCategory": {
        "type": "object",
        "properties": {
          "name_category": {
            "type": "string",
            "description": "Category: Prisma!",
            "example": "Prisma"
          }
        }
      },
      "RequestComment": {
        "type": "object",
        "properties": {
          "text_comment": {
            "type": "string",
            "description": "Comentário a ser criado",
            "example": "Qual comando para realizar uma migração no prisma ?"
          },
          "userId": {
            "type": "number",
            "description": "ID do usuário",
            "example": 1
          },
          "categoryId": {
            "type": "number",
            "description": "ID da categoria",
            "example": 1
          }
        }
      },
      "ResponseComment": {
        "type": "object",
        "properties": {
          "text_comment": {
            "type": "string",
            "description": "Comentário a ser criado",
            "example": "Qual comando para realizar uma migração no prisma ?"
          },
          "userId": {
            "type": "number",
            "description": "ID do usuário",
            "example": 2
          },
          "categoryId": {
            "type": "number",
            "description": "ID da categoria",
            "example": 2
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "Autenticação utilizando jwt",
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}